requirements.txt
ipykernel
plotly
nbformat
calplot
pandas
wordcloud
natasha
nltk
setuptools
python-dotenv
supabase